refactor: create a separate file for database functions
